requirements.txt
ultralytics          # for YOLOv8
opencv-python
numpy
pyyaml
filterpy             # for a simple SORT tracker (optional)
